Refund: cast quantity as float in JS total + controller
Yesterday's decimal-quantity fix moved the refund qty input to step=0.01
and uncapped the max, but two integer-casts downstream still clipped
3.5 to 3:
- updateRefundTotal() in view.php used parseInt(el.value), so the
  displayed refund subtotal was 3 * unit_price instead of 3.5.
- TransactionController::refund cast the posted quantity to (int),
  so even with the UI fixed, the backend would still refund 3 of 3.5.

Symptom: 350g loose tea showed a refund total closer to the 300g amount.

+++ app/controllers/TransactionController.php
@@ -131,7 +131,7 @@ class TransactionController {
         $refundItems = [];
         $refundQtys = $_POST['refund_qty'] ?? [];
         foreach ($refundQtys as $itemId => $qty) {
-            $qty = (int)$qty;
+            $qty = (float)$qty;
             if ($qty > 0) {
                 $refundItems[] = ['item_id' => (int)$itemId, 'quantity' => $qty];
             }
